blog_with_AI_bot/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
daphne==4.2.1
Django==5.2.6
django_celery_results==2.6.0
djangorestframework==3.16.1
executing==2.2.1
flower==2.0.1
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
hyperlink==21.0.0
icecream==2.1.8
idna==3.10
incremental==24.7.2
kiwisolver==1.4.9
kombu==5.5.4

msgpack==1.1.1
numpy==2.3.3
ollama==0.5.4
packaging==25.0
pillow==11.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.2
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==8.0
